requirements.txt
amqp==5.0.1
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
backcall==0.1.0
billiard==3.6.3.0
black==19.10b0
boto3==1.15.18
botocore==1.18.18
cachetools==4.1.1
celery==5.0.0
certifi==2019.3.9
cffi==1.14.1
chardet==3.0.4
Click==7.0
click-didyoumean==0.0.3
click-repl==0.1.6
coverage==4.5.3
cryptography==3.0
currency-cloud==4.4.1
Cython==0.29.21
decorator==4.4.2
deprecation==2.0.7
Django==3.1.1
django-cors-headers==3.2.1
django-dbconn-retry==0.1.5
django-dramatiq==0.5.0
django-environ==0.4.5
django-extensions==2.2.9
django-ipware==3.0.1
django-opentracing==1.1.0
django-redis==4.12.1
django-storages==1.9.1
djangorestframework==3.12.2
dramatiq==1.3.0
entrypoints==0.3
Envelopes==0.4
fin-choices-utils==1.0.4
flake8==3.7.7
flake8-coding==1.3.1
flake8-quotes==1.0.0
google-api-core==1.22.0
google-auth==1.19.2
google-cloud-core==1.3.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
gunicorn==19.9.0
html2text==2020.1.16
idna==2.8
importlib-metadata==2.0.0
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
iso3166==1.0.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.10.0
jsonfield==2.0.2
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
opentracing==2.3.0
orm-choices==1.0.0
otpauth==1.0.1
packaging==20.1
pandas==1.0.1
parso==0.6.2
pathspec==0.7.0
pdfkit==0.6.1
pexpect==4.8.0
phonenumbers==8.12.7
pickleshare==0.7.5
Pillow==7.2.0
plaid-python==4.1.0
prometheus-client==0.2.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycountry==20.7.3
pycparser==2.20
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyparsing==2.4.6
python-dateutil==2.8.0
python-http-client==3.2.7
python-magic==0.4.18
python-stdnum==1.13
pytz==2018.9
PyYAML==5.3
qrcode==6.1
raven==6.10.0
redis==3.2.0
regex==2020.1.8
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
schwifty==2020.5.3
sendgrid==6.4.3
sentry-sdk==0.19.3
six==1.12.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
toml==0.10.0
tqdm==4.48.0
traitlets==4.3.3
typed-ast==1.4.1
ua-parser==0.10.0
urllib3==1.24.1
user-agents==2.2.0
vine==5.0.0
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.11.2
XlsxWriter==1.1.5
zipp==3.3.0
